requirements.txt
python-dotenv>=1.2.1
python-telegram-bot==22.5
requests==2.32.5
pandas>=2.3.3
numpy>=2.3.4
ccxt==4.5.14